Add applicable bookstore work from last assignment
This hooks up some of the different activities and implements the
Parcelable interface in the Book and Author classes. It also moves the
targetSdkVersion up to 26 instead of 25, I was getting errors in Android
Studio at 25 and couldn't compile the app, possibly because I was
targeting a "non-current version of Android" without compatibility.