Finish the rest of the database helper
This fills out the rest of the TODO items in the database helper for the
bookstore app. Currently it doesn't actually make use of the Authors
table, this is something that will need to be done.

To finish the bookstore app all that is left is hooking up the database
to the MainActivity and getting all the activities properly interacting
with the database. Then actually splitting the authors out into the
other table.