Add database helper to MainActivity
This hooks up the database helper to the MainActivity and adds the
ContextMenu to hook into the ViewBookActivity. Currently the books in
the cart appear with light text on a light background, not sure why.

Also, still TODO is making the database helper save authors in a
separate table from books. After that the bookstore app should be pretty
much done.