Fix cart text color
For some reason the text of the cart_row items in the MainActivity
displayed as light on a light background. I think it may be some sort of
bug with the Theme.Material.Light.DarkActionBar (which seems to be a
light theme with a dark actionbar). The ListView items seem to think
it's a dark theme. Either way forcing the text to
color/primary_text_light resolves the issue.