Changed bookstore context menu to a CAB
This changes the long press ContextMenu into a Contextual Action Bar, it
only supports single item selection because multi-select seems to need a
certain adapter for the ListView and this already has a
SimpleCursorAdapter and I don't want to mess that up.

This also makes a small adjustment to the Author class so that it only
saves a String of length one in the middleInitial field (so actually
saving an initial instead of a full name).