Start separating tables in bookstore app
This starts all the work involved with separating the tables (books and
authors) in the bookstore app. Most of the work database side is done,
but the rest of the app doesn't work with it properly yet. The cart
doesn't properly get the authors so the string is null, which crashes
the ViewBookActivity.