Add chat client and finish up chat server
This adds a chat client app, which is a copy of the assignment 2 client
with some small tweaks to how the message is sent. This is needed
because the assignment 3 chat server expects to receive messages
formatted as: "name:timestamp:message".

This also finishes hooking up the chat server, there were some small
additions to get the ids of peers in the ViewPeerActivity.

The issue I saw with the bookstore app where text is light on a light
background is back, and I tried the same solution but it didn't work
because I wasn't actually using the layout that I edited. The text is
being displayed in default android layouts so I'm not sure if I can
force it's color or not.

Other than that the assignment is finished, all that is needed is a
recording of the chat app working.