Finish up assignment 3 chat apps
This adds final tweaks to the chat server. It makes use of the message
view in order to fix the white on light background text issue, and it
updates the database insert function for peers so that it updates
entries if a duplicate is found.

This also adds the recording for the chat apps and a README report on
the submission.