Enable saving and deleting from content provider
This hooks up the content provider to the MainActivity so that books
returned from the AddBookActivity update the content provider. It also
registers the cursor in the LoaderManager to listen for updates to the
books URI so that the adapter automatically updates.

This comments out all the lines dealing with authors to get it running
for now. The BookProvider will need to be updated to insert authors into
the authors table and return queries from a join of the two tables in
order to return book authors.

After that the only major thing left will be figuring out why the CAB
doesn't work.