Add authors to content provider
This updates the assignment 4 bookstore app to save authors. They are
passed into the content provider in the ContentValues object that the
book is passed in (using the BookContract putAuthors function). Then in
the content provider they are pulled out and inserted into the authors
table.

All content provider query functions have been updated to query from a
join of the books and authors tables, with a selection and a GROUPBY
statement that will fill in the BookContract.AUTHORS synthetic column.