Fix author list text color and add checkout Toast
This adds a layout for the author list in the ViewBookActivity so that I
can force it's text color to not be white on a light background.

This also passes the number of books in the cart to the CheckoutActivity
so that it can display it's Toast message stating how many books were
"ordered".

All that is left for the bookstore app is to fix the CAB and get
multi-select deleting working.