Add applicable work from assignment 4
This goes through and adds in the work already done for this assignment
that was done in the last one. This includes the entities, content
provider, and most of the UI. The MainActivity will still need to be
filled out with asynchronous entity managers and other callbacks, which
themselves still need to be fully implemented.

This also updates the repository README to add assignment 5 and use
bullet points instead of numbers, as I think the numbers look redundant
since the assignments are numbered anyway.