Fix messages not being added to content provider
The sending and receiving of the messages was working fine, but since
they were being received in a background Service they were not able to
be added to the Content Provider because the Entity Managers required an
Activity Context.

This fixes that by getting a synchronous Content Resolver in the Service
(which shouldn't matter because the service is in the background anyway)
and directly using that to query and insert into the Content Provider.

The Entity Manager code has been left in place but commented out.