Fix crashes due to permissions and incorrect IP
This fixes a few more crashes. First off the app needs WAKE_LOCK
permission or it will crash on registering. The ContentProvider was not
declared in the manifest, causing an error on startup. The
RequestService was not initializing it's super class properly, causing a
crash on registering. Also the base_uri IP address was pointing the
emulator at itself, the emulator is able to reach the development host
on 10.0.2.2.