Fix SettingsActivity not displaying settings
The SettingsActivity previously didn't display the username or UUID,
this changes the xml to use the same Settings keys as Settings.java and
also populates the preferences with their values.

In order not to have a literal string duplicated in the code the keys in
Settings.java that were needed were made public. As these are just
strings it should have no consequences elsewhere.