Fix ContentProvider issues
The app kept crashing when sending a message because it couldn't add it
to the ContentProvider. At first this was because longitude and latitude
were null, then it was because the Peer foreign key reference failed.

After trying to add Peers to the database I ended up skipping the
AsyncContentResolver like I did in assignment 6. When trying to add
Peers the executeSimpleQuery function failed because the context could
not be cast to an Activity.

This now adds Messages and Peers to the local database synchronously,
but like in assignment 6 we are already off the UI thread so this
shouldn't matter too much.